backend/requirements.txt
asgiref==3.10.0
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
cryptography==46.0.3
defusedxml==0.7.1
distlib==0.4.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
filelock==3.20.0
flake8==7.3.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python3-openid==3.2.0
pytokens==0.2.0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
virtualenv==20.35.3
wheel==0.45.1
